py/server/requirements.txt
flask>=2.3.0
requests>=2.31.0
pillow>=10.0.0
torch>=2.1.0; platform_system != 'Windows' or platform_machine == 'x86_64'
torchvision>=0.16.0; platform_system != 'Windows' or platform_machine == 'x86_64'
opencv-python>=4.8.0
scipy>=1.11.3
numpy>=1.24.0
matplotlib>=3.7.3
kafka-python>=2.0.2
redis>=5.0.1
